refactor(importer): split scanner.go god-file into focused files
Pure mechanical split of the ~3,300-LOC scanner.go into three cohesive
files within the same importer package. Whole functions/types moved
verbatim; no behavior, signature, logic, comment, or name changes. Only
per-file import blocks adjusted to compile.

- scanner.go: spine (Scanner struct, NewScanner, With* wiring,
  CheckDownloads, recovery, ScanLibrary/writeScanResult,
  tryImportInternal core, shared helpers).
- scanner_poll.go: per-client poll/reconcile (checkSABnzbdDownloads,
  checkNZBGetDownloads, checkTransmissionDownloads,
  checkQbittorrentDownloads), tryImport* wrappers, and torrent-file
  resolution/matching helpers.
- scanner_handoff.go: post-import handoffs (pushToCWA, pushToCalibre,
  pushCalibreAdd, pushToABS) and the #997 drop-folder code.

Top-level func/type name set unchanged (92 before, 92 after, no
diff). build + vet + gofmt + importer tests + golangci-lint all pass.
